// src/core-client/generated/PlaybackCommandDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type PlaybackCommandDto = { "type": "setPaused", paused: boolean, } | { "type": "seekAbsolute", position: number, } | { "type": "seekRelative", seconds: number, } | { "type": "setVolume", volume: number, } | { "type": "setMuted", muted: boolean, } | { "type": "setSpeed", speed: number, } | { "type": "stop" };